Fixed warning about phpunit suite
