+++ phpunit.xml.dist
@@ -5,14 +5,13 @@
          colors="true">
 
     <testsuites>
-        <testsuite name="all">
-            <directory>tests/</directory>
+        <testsuite name="mysql">
+            <directory>tests/Mysql/</directory>
             <exclude>tests/_files/*</exclude>
         </testsuite>
         <testsuite name="unit">
-            <directory>tests/</directory>
+            <directory>tests/Unit/</directory>
             <exclude>tests/_files/*</exclude>
-            <exclude>tests/Operation/OperationsMySQLTest.php</exclude>
         </testsuite>
     </testsuites>
