+++ tests/Unit/DataSet/ReplacementTableTest.php
@@ -9,70 +9,245 @@
  * file that was distributed with this source code.
  */
 
-namespace PHPUnit\DbUnit\Tests\DataSet;
+namespace PHPUnit\DbUnit\Tests\Unit\DataSet;
 
 use PHPUnit\DbUnit\DataSet\DefaultTable;
 use PHPUnit\DbUnit\DataSet\DefaultTableMetadata;
 use PHPUnit\DbUnit\DataSet\ITable;
 use PHPUnit\DbUnit\DataSet\ITableMetadata;
-use PHPUnit\DbUnit\DataSet\QueryTable;
+use PHPUnit\DbUnit\DataSet\ReplacementTable;
+use PHPUnit\DbUnit\TestCase;
 use PHPUnit\Framework\MockObject\MockObject;
-use PHPUnit\Framework\TestCase;
 
-class AbstractTableTest extends TestCase
+class ReplacementTableTest extends \PHPUnit\Framework\TestCase
 {
     /**
-     * @var QueryTable
+     * @var DefaultTable
      */
-    protected $table;
+    protected $startingTable;
 
     protected function setUp(): void
     {
         $tableMetaData = new DefaultTableMetadata(
-            'table',
-            ['id', 'column1']
+            'table1',
+            ['table1_id', 'column1', 'column2', 'column3', 'column4']
         );
 
-        $this->table = new DefaultTable($tableMetaData);
+        $table = new DefaultTable($tableMetaData);
 
-        $this->table->addRow([
-            'id' => 1,
-            'column1' => 'randomValue',
+        $table->addRow([
+            'table1_id' => 1,
+            'column1' => 'My name is %%%name%%%',
+            'column2' => 200,
+            'column3' => 34.64,
+            'column4' => 'yghkf;a  hahfg8ja h;',
         ]);
+        $table->addRow([
+            'table1_id' => 2,
+            'column1' => 'hk;afg',
+            'column2' => 654,
+            'column3' => 46.54,
+            'column4' => '24rwehhads',
+        ]);
+        $table->addRow([
+            'table1_id' => 3,
+            'column1' => 'ha;gyt',
+            'column2' => 462,
+            'column3' => '[NULL] not really',
+            'column4' => '[NULL]',
+        ]);
+
+        $this->startingTable = $table;
     }
 
-    /**
-     * @param array $row
-     * @param bool $exists
-     *
-     * @dataProvider providerTableContainsRow
-     */
-    public function testTableContainsRow($row, $exists): void
+    public function testNoReplacement(): void
     {
-        $result = $this->table->assertContainsRow($row);
-        $this->assertEquals($exists, $result);
+        TestCase::assertTablesEqual(
+            $this->startingTable,
+            new ReplacementTable($this->startingTable)
+        );
     }
 
-    public static function providerTableContainsRow(): array
+    public function testFullReplacement(): void
     {
-        return [
-            [[
-                'id' => 1,
-                'column1' => 'randomValue',
-            ], true],
-            [[
-                'id' => 1,
-                'column1' => 'notExistingValue',
-            ], false],
-        ];
+        $tableMetaData = new DefaultTableMetadata(
+            'table1',
+            ['table1_id', 'column1', 'column2', 'column3', 'column4']
+        );
+
+        $table = new DefaultTable($tableMetaData);
+
+        $table->addRow([
+            'table1_id' => 1,
+            'column1' => 'My name is %%%name%%%',
+            'column2' => 200,
+            'column3' => 34.64,
+            'column4' => 'yghkf;a  hahfg8ja h;',
+        ]);
+        $table->addRow([
+            'table1_id' => 2,
+            'column1' => 'hk;afg',
+            'column2' => 654,
+            'column3' => 46.54,
+            'column4' => '24rwehhads',
+        ]);
+        $table->addRow([
+            'table1_id' => 3,
+            'column1' => 'ha;gyt',
+            'column2' => 462,
+            'column3' => '[NULL] not really',
+            'column4' => null,
+        ]);
+
+        $actual = new ReplacementTable($this->startingTable);
+        $actual->addFullReplacement('[NULL]', null);
+
+        TestCase::assertTablesEqual($table, $actual);
+    }
+
+    public function testSubStrReplacement(): void
+    {
+        $tableMetaData = new DefaultTableMetadata(
+            'table1',
+            ['table1_id', 'column1', 'column2', 'column3', 'column4']
+        );
+
+        $table = new DefaultTable($tableMetaData);
+
+        $table->addRow([
+            'table1_id' => 1,
+            'column1' => 'My name is Mike Lively',
+            'column2' => 200,
+            'column3' => 34.64,
+            'column4' => 'yghkf;a  hahfg8ja h;',
+        ]);
+        $table->addRow([
+            'table1_id' => 2,
+            'column1' => 'hk;afg',
+            'column2' => 654,
+            'column3' => 46.54,
+            'column4' => '24rwehhads',
+        ]);
+        $table->addRow([
+            'table1_id' => 3,
+            'column1' => 'ha;gyt',
+            'column2' => 462,
+            'column3' => '[NULL] not really',
+            'column4' => '[NULL]',
+        ]);
+
+        $actual = new ReplacementTable($this->startingTable);
+        $actual->addSubStrReplacement('%%%name%%%', 'Mike Lively');
+
+        TestCase::assertTablesEqual($table, $actual);
+    }
+
+    public function testConstructorReplacements(): void
+    {
+        $tableMetaData = new DefaultTableMetadata(
+            'table1',
+            ['table1_id', 'column1', 'column2', 'column3', 'column4']
+        );
+
+        $table = new DefaultTable($tableMetaData);
+
+        $table->addRow([
+            'table1_id' => 1,
+            'column1' => 'My name is Mike Lively',
+            'column2' => 200,
+            'column3' => 34.64,
+            'column4' => 'yghkf;a  hahfg8ja h;',
+        ]);
+        $table->addRow([
+            'table1_id' => 2,
+            'column1' => 'hk;afg',
+            'column2' => 654,
+            'column3' => 46.54,
+            'column4' => '24rwehhads',
+        ]);
+        $table->addRow([
+            'table1_id' => 3,
+            'column1' => 'ha;gyt',
+            'column2' => 462,
+            'column3' => '[NULL] not really',
+            'column4' => null,
+        ]);
+
+        $actual = new ReplacementTable(
+            $this->startingTable,
+            [
+                '[NULL]' => null,
+            ],
+            [
+                '%%%name%%%' => 'Mike Lively',
+            ]
+        );
+
+        TestCase::assertTablesEqual($table, $actual);
+    }
+
+    public function testGetRow(): void
+    {
+        $actual = new ReplacementTable(
+            $this->startingTable,
+            [
+                '[NULL]' => null,
+            ],
+            [
+                '%%%name%%%' => 'Mike Lively',
+            ]
+        );
+
+        $this->assertEquals(
+            [
+                'table1_id' => 1,
+                'column1' => 'My name is Mike Lively',
+                'column2' => 200,
+                'column3' => 34.64,
+                'column4' => 'yghkf;a  hahfg8ja h;',
+            ],
+            $actual->getRow(0)
+        );
+
+        $this->assertEquals(
+            [
+                'table1_id' => 3,
+                'column1' => 'ha;gyt',
+                'column2' => 462,
+                'column3' => '[NULL] not really',
+                'column4' => null,
+            ],
+            $actual->getRow(2)
+        );
+    }
+
+    public function testGetValue(): void
+    {
+        $actual = new ReplacementTable(
+            $this->startingTable,
+            [
+                '[NULL]' => null,
+            ],
+            [
+                '%%%name%%%' => 'Mike Lively',
+            ]
+        );
+
+        $this->assertNull($actual->getValue(2, 'column4'));
+        $this->assertEquals('My name is Mike Lively', $actual->getValue(0, 'column1'));
     }
 
     public function testMatchesWithNonMatchingMetaData(): void
     {
         $tableMetaData = $this->createMock(ITableMetadata::class);
         $otherMetaData = $this->createMock(ITableMetadata::class);
-
+        $table = $this->createMock(ITable::class);
         $otherTable = $this->createMock(ITable::class);
+
+        $table->expects($this->once())
+            ->method('getTableMetaData')
+            ->willReturn($tableMetaData);
+
         $otherTable->expects($this->once())
             ->method('getTableMetaData')
             ->willReturn($otherMetaData);
@@ -82,21 +257,26 @@ class AbstractTableTest extends TestCase
             ->with($otherMetaData)
             ->willReturn(false);
 
-        $table = new DefaultTable($tableMetaData);
-        $this->assertFalse($table->matches($otherTable));
+        $replacementTable = new ReplacementTable($table);
+        $this->assertFalse($replacementTable->matches($otherTable));
     }
 
     public function testMatchesWithNonMatchingRowCount(): void
     {
         $tableMetaData = $this->createMock(ITableMetadata::class);
         $otherMetaData = $this->createMock(ITableMetadata::class);
+        $table = $this->createMock(ITable::class);
         $otherTable = $this->createMock(ITable::class);
 
-        /** @var MockObject|DefaultTable $table */
-        $table = $this->getMockBuilder(DefaultTable::class)
-            ->setConstructorArgs([$tableMetaData])
+        /** @var MockObject|ReplacementTable $replacementTable */
+        $replacementTable = $this->getMockBuilder(ReplacementTable::class)
+            ->setConstructorArgs([$table])
             ->onlyMethods(['getRowCount'])
             ->getMock();
+
+        $table->expects($this->once())
+            ->method('getTableMetaData')
+            ->willReturn($tableMetaData);
 
         $otherTable->expects($this->once())
             ->method('getTableMetaData')
@@ -110,17 +290,17 @@ class AbstractTableTest extends TestCase
             ->with($otherMetaData)
             ->willReturn(true);
 
-        $table->expects($this->once())
+        $replacementTable->expects($this->once())
             ->method('getRowCount')
             ->willReturn(1);
 
-        $this->assertFalse($table->matches($otherTable));
+        $this->assertFalse($replacementTable->matches($otherTable));
     }
 
     /**
      * @param array $tableColumnValues
      * @param array $otherColumnValues
-     * @param bool $matches
+     * @param bool  $matches
      *
      * @dataProvider providerMatchesWithColumnValueComparisons
      */
@@ -128,13 +308,12 @@ class AbstractTableTest extends TestCase
     {
         $tableMetaData = $this->createMock(ITableMetadata::class);
         $otherMetaData = $this->createMock(ITableMetadata::class);
+        $table = $this->createMock(ITable::class);
         $otherTable = $this->createMock(ITable::class);
 
-        /** @var MockObject|DefaultTable $table */
-        $table = $this->getMockBuilder(DefaultTable::class)
-            ->setConstructorArgs([$tableMetaData])
-            ->onlyMethods(['getRowCount', 'getValue'])
-            ->getMock();
+        $table->expects($this->once())
+            ->method('getTableMetaData')
+            ->willReturn($tableMetaData);
 
         $otherTable->expects($this->once())
             ->method('getTableMetaData')
@@ -151,7 +330,13 @@ class AbstractTableTest extends TestCase
             ->with($otherMetaData)
             ->willReturn(true);
 
-        $table
+        /** @var MockObject|ReplacementTable $replacementTable */
+        $replacementTable = $this->getMockBuilder(ReplacementTable::class)
+            ->setConstructorArgs([$table])
+            ->onlyMethods(['getRowCount', 'getValue'])
+            ->getMock();
+
+        $replacementTable
             ->method('getRowCount')
             ->willReturn(\count($tableColumnValues));
 
@@ -164,14 +349,14 @@ class AbstractTableTest extends TestCase
                 $otherMap[] = [$rowIndex, $columnName, $otherColumnValues[$rowIndex][$columnName]];
             }
         }
-        $table
+        $replacementTable
             ->method('getValue')
             ->willReturnMap($tableMap);
         $otherTable
             ->method('getValue')
             ->willReturnMap($otherMap);
 
-        $this->assertSame($matches, $table->matches($otherTable));
+        $this->assertSame($matches, $replacementTable->matches($otherTable));
     }
 
     public static function providerMatchesWithColumnValueComparisons(): array
